requirements.txt
streamlit>=1.30.0
Pillow>=9.0.0
simplekml>=1.3.0
pandas>=1.5.0
PyMuPDF>=1.23.0
numpy>=1.24.0
